api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pyyaml==6.0.1
opencv-python==4.11.0.86
numpy==2.1.2
